[cxx-interop] [nfc] `llvm::Optional<AnyFunctionRef>` -> `std::unique_ptr<AnyFunctionRef>`.
`AnyFunctionRef` is a forward declared class in `SILDeclRef.h` this is requred for dependency /layering reasons. However, we cannot get the alignement, etc. of a forward declared class, so we can't use it with `llvm::Optional`. It must only be a pointer.

(This is needed to be able to import `SILDeclRef.h` with C++ interop.)

This also helps us follows the general rule that C++ headers should be self-contained.