[moveOnly] Teach loadable by address to just delete move_value that it sees when it is rewriting code.
For now I am preserving move_value to the end of the pipeline. This just
prevents LoadableByAddress from needing a move_value_addr. We can add that if we
need to though. For now we just RAUW the move_value as we convert the function
argument from its object to its addr form.

TLDR: This just simplifies loadable by address.

rdar://83957088