cmake: add a cache for Windows ARM64 toolchains
This adds a CMake cache for the Windows ARM64 toolchain.  It is possible
to build and run the toolchain on ARM64, though we have not had regular
releases of the toolchain.  This adds the cache used to generate the
toolchain in the hopes that it is useful for others and that we can soon
get more regular ARM64 toolchain builds.