improve robustness of dist actor test
We were seeing flaky runs due to the hard-to-determine order and point at which
the actor instances were being deallocated in the old version of the test. In addition,
all of the actor addresses were the same, so moving the point of deallocation would
give us a less rigorous test.

This patch puts all of the instances in an array so that we can rely on CHECK-DAG
to check for the deallocations in any order, as a group at the end of the test. The
deallocations are now identifiable using a simple counter for the addresses.

resolves rdar://84574311