fix the isolation of the `self` param in ctors and dtors
The isolation of `self` was missing on its ParamDecl of ctors
of an actor, leading to situations where closures that should
inherit the isolation of the ctor, not actually inheriting it
because they only inspect the ParamDecl.

This patch makes it so that both inits and deinits also
get their `self` isolation by making a query in the
type checker.

resolves rdar://84682865