Add `_getMainExecutor`
The `getMainExecutor` was declared internal so it doesn't show up in the
swift interface files. The definition exists though. The missing
declaration causes the compiler to crash while compiling programs with
async-main since it can't find the declaration.

This patch changes it to conjure up the appropriate decl if it can't
find it.