Reinforce the Sendable requirement on sync actor inits
With the change I made to say that synchronous actor inits
that are otherwise not isolated to anything else should be `nonisolated`,
we lost the Sendable checking on its parameters.

This patch brings that back by enforcing a CrossActorSelf restriction
for accesses to such decls, despite them having `nonisolated` isolation.