GSB: Loosen up verifyGenericSignature()
We used to assert if a generic requirement of a signature could
be proved from the signature with the requirement removed.

However, in some rare cases minimization can *add* new requirements
as well as remove them.

This is going to come up even more with the RequirementMachine,
where for example the following signature is accepted:

    protocol Q { associatedtype A : P }
    protocol P { associatedtype B }
    <T where T.A : Q, T.A.B == T>

and minimized as follows:

    <T where T : P, T.A : Q, T.A.B == T>