Relax warnings about interior ManagedBuffer pointers
In order to be useful for large allocations, `ManagedBuffer` must guarantee the following:
- Its contents are never moved or reallocated. (Elements or header properties might be nontrivial to copy or move.)
- Its header is always located close to its first element. (In practice, ManagedBuffer lays out its storage as if it were a struct of `{Header, (Element, Element ...)}`.)

The documentation for the `withUnsafePointer...` methods creates serious doubt in the programmer’s mind about the validity of inferring these behaviors. Since the documentation implies that the validity of the `header` and `elements` pointers is tied to the duration of the call and not the lifetime of the object, that makes `ManagedBuffer` unsuitable for storing locks or atomic variables in either its `header` *or* `elements`, since these data types rely on the stability of their address in memory. Common wisdom seems to have arisen that it is in fact safe to rely on the location of `elements`, but not of `header`. This results in the unnecessarily complicated recommendation to allocate a heterogenous `ManagedBuffer<UInt8>` and manage its internal layout manually, rather than storing any locks or atomic variables in `header`.

Rather than continue to sow doubt, let’s just publicly commit to the stability of any given `ManagedBuffer`’s location and layout in memory.